+++ rust/dbn/src/python.rs
@@ -1351,7 +1351,11 @@ impl ErrorMsg {
     }
 
     fn __repr__(&self) -> String {
-        format!("{self:?}")
+        if let Ok(err_msg) = self.err() {
+            format!("ErrorMsg {{ hd: {:?}, err: '{}' }}", self.hd, err_msg)
+        } else {
+            format!("ErrorMsg {{ hd: {:?}, err: '{:?}' }}", self.hd, self.err)
+        }
     }
 
     #[getter]
@@ -1497,7 +1501,11 @@ impl SystemMsg {
     }
 
     fn __repr__(&self) -> String {
-        format!("{self:?}")
+        if let Ok(sys_msg) = self.msg() {
+            format!("SystemMsg {{ hd: {:?}, msg: '{}' }}", self.hd, sys_msg)
+        } else {
+            format!("SystemMsg {{ hd: {:?}, msg: '{:?}' }}", self.hd, self.msg)
+        }
     }
 
     #[getter]